Extract formatting, compression, and migration into base class, making it abstract
